Use 1MB read/write size and drop the need for blockdev
The code *used* to read and write using whatever block size the
drive reported as its physical block size.

This caused the read and write throughput to be pretty slow,
such that the program took about 1.5 times as long to run as
badblocks -wvs.

Switching to writing in fixed 1MB chunks instead (as badblocks
seems to do) has dramatically sped things up.  The program now
finishes in about 75% of the time that badblocks -wvs does.

For example, `badblocks -wvs /dev/zd48` takes about 35 seconds
to run.  The previous code for this utility used to take about
54 seconds for the same device.

But using 1MB read/write size now has the program completing in
about 25 seconds. :)